% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_convert.R
\name{batch_convert}
\alias{batch_convert}
\title{Convert all files of a certain type to SAGA format}
\usage{
batch_convert(signature)
}
\arguments{
\item{signature}{A file extension as a string. It is recommended that this string end in a `$` character.}
}
\value{
Nothing. Files with the same name as the inputs (but with different file extensions) will be available in the current working directory.
}
\description{
Given a file extension, this function converts all files with that file extension in the current working directory into SAGA format.
}
