% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_running.R
\name{check_running}
\alias{check_running}
\title{Check whether a valid GRASS session is running}
\usage{
check_running()
}
\value{
A logical. The logical indicates whether a valid GRASS session is currently running.
}
\description{
This function is mostly used internally by other functions in the package. However, users may call this function to check whether they have correctly established a GRASS session prior to using the other functions in the package.
}
\examples{

check_running()

}
