% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_mask.R
\name{clear_mask}
\alias{clear_mask}
\title{Clear current raster mask}
\usage{
clear_mask()
}
\description{
This function has no parameters. It can be used to clear an existing raster mask.
}
\examples{
if(check_running()) clear_mask()
}
