% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_iFLO_weights.R
\name{compute_iFLO_weights}
\alias{compute_iFLO_weights}
\title{Compute iFLO weights}
\usage{
compute_iFLO_weights(
  pour_point,
  watershed,
  null_streams,
  flow_dir,
  out_flow_length,
  out_iFLO,
  out_iFLO_no_stream,
  idwp = -1,
  remove_streams = FALSE,
  ...
)
}
\arguments{
\item{pour_point}{Pour point raster containing a single pour point (i.e., the outlet).}

\item{watershed}{Watershed raster to use as a mask for the flow-path calculations.}

\item{null_streams}{A streams raster with NoData for the stream cells and 1s everywhere else}

\item{flow_dir}{A flow direction raster.}

\item{out_flow_length}{Name of the output flow length raster.}

\item{out_iFLO}{Name of the output weights raster.}

\item{out_iFLO_no_stream}{Name of the output weights raster excluding cells on the stream line (ignored inf \code{remove_streams = FALSE}).}

\item{idwp}{An inverse distance weighting power. This should be negative. The value \code{idwp = -1} is the default.}

\item{remove_streams}{A logical indicating whether cells corresponding to the stream line should be removed from the weights raster. Defaults to \code{FALSE}.}
}
\value{
Nothing.
}
\description{
Compute an iFLO weight raster outside of the \code{compute_metrics()} function.
}
