% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{compute_metrics}
\alias{compute_metrics}
\title{Compute spatially explicit watershed attributes for survey sites on streams}
\usage{
compute_metrics(
  metrics = c("lumped", "iFLO", "iFLS", "HAiFLO", "HAiFLS"),
  landuse,
  sites,
  watersheds,
  flow_dir,
  flow_acc,
  streams,
  idwp = -1,
  max_memory = 300
)
}
\arguments{
\item{metrics}{A character vector. This vector specifies which metric(s) should be calculated. Your options are lumped, iFLO, iFLS, iEDO, iEDS, HAiFLO and HAiFLS. The default is to calculate all except for lumped, iEDO and iEDS.}

\item{landuse}{Names of binary landuse or landcover rasters in the current GRASS mapset for which spatially explicit watershed metrics should be computed. Binary means land use cells are coded 1 and all other cells are given a value of 0.}

\item{sites}{A vector data set of survey sites in the current GRASS mapset.}

\item{watersheds}{A vector of watershed raster names in the current GRASS mapset.}

\item{flow_dir}{Name of a flow direction raster produced}

\item{flow_acc}{Name of a flow accumulation raster produced by \code{derive_flow} in the current GRASS mapset.}

\item{streams}{Name of a streams raster in the current GRASS mapset. Optional if you are not asking for the iFLS, iEDS, and/or HAiFLS metrics.}

\item{idwp}{The inverse distance weighting parameter. Default is \code{-1}.}

\item{max_memory}{Max memory used in memory swap mode (MB). Defaults to \code{300}.}

\item{lessmem}{A logical indicating whether to use the less memory modified watershed module. Defaults to \code{FALSE}.}
}
\value{
A \code{sf} object, which is a table with columns for each combination of land use and metric type.
}
\description{
Workhorse function for \code{rdwplus}. This function computes the spatially explicit landuse metrics in IDW-Plus (Peterson and Pearse, 2017).
}
\examples{
# Will only run if GRASS is running
if(check_running()){
}
}
\references{
Peterson, E.E. & Pearse, A.R. (2017). IDW-Plus: An ArcGIS toolset for calculating spatially explicit watershed attributes for survey sites. \emph{JAWRA}, \emph{53}(5), 1241-1249.
}
