% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_sgrd.R
\name{convert_to_sgrd}
\alias{convert_to_sgrd}
\title{Facilitate the conversion of files to SAGA format}
\usage{
convert_to_sgrd(in_file, out_file)
}
\arguments{
\item{in_file}{A vector of strings, which are file paths to existing files.}

\item{out_file}{A vector of strings, which are file paths to the output files.}
}
\value{
Nothing. The files will be saved wherever is indicated in out_file.
}
\description{
The SAGA modules called in this package require files to be in SAGA format, so this function calls GDAL to convert any non-SAGA files to the required format.
}
