% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_to_raster.R
\name{coord_to_raster}
\alias{coord_to_raster}
\title{Turn coordinates of outlets into rasters}
\usage{
coord_to_raster(outlet, out, overwrite = FALSE)
}
\arguments{
\item{outlet}{A single pair of Easting, Northing or long, lat coordinates as a numeric vector.}

\item{out}{The file name of the output outlet raster in the current GRASS mapset.}

\item{overwrite}{Whether the output files should be allowed to overwrite existing files. Defaults to \code{FALSE}.}
}
\value{
Nothing.
}
\description{
Given a set of coordinates in space (x, y), this function will return a rasterised version of that point in space.
}
