% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_streams.R
\name{derive_streams}
\alias{derive_streams}
\title{Extract streams from a flow accumulation raster}
\usage{
derive_streams(
  dem,
  flow_acc,
  out_rast,
  out_vect,
  min_acc = 1000,
  min_length = 0,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{dem}{Name of an elevation raster in the current GRASS mapset.}

\item{flow_acc}{Name of a flow accumulation raster in the current GRASS mapset.}

\item{out_rast}{Name of the output raster dataset of stream lines. File extensions should not matter.}

\item{out_vect}{Name of the output vector dataset of stream lines. Should be WITHOUT .shp extension.}

\item{min_acc}{The minimum accumulation value that a cell needs to be classified as a stream. Defaults to \code{1000}.}

\item{min_length}{The minimum length of a stream segment in cells. Defaults to \code{0}.}

\item{overwrite}{A logical indicating whether the output should be allowed to overwrite existing files. Defaults to \code{FALSE}.}

\item{...}{Additional arguments to \code{r.stream.extract}.}
}
\value{
Nothing. A vector dataset with the name \code{basename(out)} will appear in the current GRASS mapset.
}
\description{
Derive a raster and a vector layer of stream lines from a flow accumulation raster.
}
\examples{
 
# Will only run if GRASS is running
if(check_running()){
}
}
