% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_sinks.R
\name{fill_sinks}
\alias{fill_sinks}
\title{Fill sinks in a digital elevation model (DEM)}
\usage{
fill_sinks(dem, out, ...)
}
\arguments{
\item{dem}{File path to a digital elevation model, in SAGA format (.sgrd extension)}

\item{out}{Name of the file to be written out, INCLUDING a .sgrd extension. This will be the filled DEM.}

\item{...}{Optional additional parameters to \code{rsaga.fill.sinks}. Note that the \code{method} is hard-coded to \code{'xxl.wang.liu.2006'}.}
}
\value{
Nothing. A file with the name \code{out} will be created in the current working directory.
}
\description{
A sink is a depression in a DEM. Water flows into these depressions but not out of them. These depressions, although often real features of landscapes, are problematic for flow direction and accumulation algorithms. Therefore, it is common practice to remove these depressions. This function removes depressions (sinks) in a DEM.
}
