% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distance.R
\name{get_distance}
\alias{get_distance}
\title{Compute Euclidean distance to a survey site or stream line within a watershed}
\usage{
get_distance(background, target, out_file)
}
\arguments{
\item{background}{A raster of the watershed outline. This can also be a file path to an existing raster.}

\item{target}{A raster of the survey site or stream lines. This can also be a file path to an existing raster.}

\item{out_file}{File path for the result to be written.}
}
\value{
Nothing. A file with the name and location \code{out_file} will be created.
}
\description{
This function is needed to compute Euclidean distance from a feature of interest in a watershed raster.
}
