% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flowaccumulation.R
\name{get_flowaccumulation}
\alias{get_flowaccumulation}
\title{Derive a flow accumulation grid from a digital elevation model (DEM) under the D8 flow direction algorithm}
\usage{
get_flowaccumulation(dem, out, ...)
}
\arguments{
\item{dem}{A file path to the hydrologically corrected DEM, with .sgrd extension.}

\item{out}{A file name for the output flow accumulation grid, with .sgrd extension.}

\item{...}{Optional arguments to \code{rsaga.geoprocessor}.}
}
\value{
Nothing. A file with the name \code{out} will be written to SAGA's current workspace.
}
\description{
Given a (hydrologically corrected, see \code{\link{fill_sinks}}) DEM, this function produces a flow accumulation grid which shows the upstream area that flows into each cell in the DEM.
}
