% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SAGA_distance.R
\name{get_SAGA_distance}
\alias{get_SAGA_distance}
\title{Compute Euclidean distance to a survey site or stream line within a watershed}
\usage{
get_SAGA_distance(background, target, out_file)
}
\arguments{
\item{background}{File path to an existing raster, which has non NoData cells for the watershed over which distances are being computed.}

\item{target}{File path to an existing raster, from which distances will be computed.}

\item{out_file}{File path for the result to be written.}
}
\value{
Nothing. A file with the name and location \code{out_file} will be created.
}
\description{
This function is needed to compute Euclidean distance from a feature of interest in a watershed raster. The computations here are performed in the background by SAGA. This is required to manage memory correctly even for small rasters (~10s of MB).
}
