% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_watershed.R
\name{get_watershed}
\alias{get_watershed}
\title{Delineate watershed for a survey site}
\usage{
get_watershed(sites, i, dem, out, R = TRUE)
}
\arguments{
\item{sites}{A \code{SpatialPoints*} object.}

\item{i}{An integer which indexes one row of the sites' attribute table.}

\item{dem}{Pointer to a hydrologically-corrected digital elevation model.}

\item{out}{The name of the output file, with NO file extension.}

\item{R}{A logical indicating whether a RasterLayer should be returned in R. Defaults to \code{TRUE}.}
}
\value{
A SAGA raster with basename \code{out}. Possibly also a \code{RasterLayer} object.
}
\description{
This function delineates a watershed around the \code{i}th site from a set of survey sites.
}
