% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_extensions.R
\name{install_extensions}
\alias{install_extensions}
\title{Install required extension(s)}
\usage{
install_extensions()
}
\value{
Nothing.
}
\description{
Some functions in the \code{rdwplus} package rely on GRASS extensions that need to be installed prior to use. This function installs those extensions.
}
\details{
This function has no arguments. Simply run it and it will install a pre-set list of GRASS extensions.

Currently, the GRASS extension required are \code{r.stream.snap}, \code{r.stream.distance}, and \code{r.wateroutlet.lessmem}.
}
\examples{
# Will only run if GRASS is running
if(check_running()){
    install_extensions()
}
}
