% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_layer.R
\name{plot_layer}
\alias{plot_layer}
\title{A function to plot a raster from a file name}
\usage{
plot_layer(x, ...)
}
\arguments{
\item{x}{A file path to a raster stored on disk. Can also be a \code{Raster*} object.}

\item{...}{Additional arguments to \code{plot}.}
}
\value{
Nothing.
}
\description{
Given a path to a file raster, this function will plot it as a \code{RasterLayer} object.
}
\examples{
\dontrun{
# Load data set file path
dem <- system.file("extdata", "dem.tif", package = "rdwplus")
# Plot
plot_layer(dem)
}
}
