% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast_calc.R
\name{rast_calc}
\alias{rast_calc}
\title{Raster calculator (wrapper for "r.mapcalc")}
\usage{
rast_calc(ex, overwrite = TRUE)
}
\arguments{
\item{ex}{A valid raster calculator expression for GRASS.}

\item{overwrite}{Defaults to \code{TRUE}.}
}
\value{
Nothing.
}
\description{
Raster calculator (wrapper for "r.mapcalc")
}
