% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_to_mapset.R
\name{raster_to_mapset}
\alias{raster_to_mapset}
\title{Import rasters into GRASS mapset}
\usage{
raster_to_mapset(rasters)
}
\arguments{
\item{rasters}{A character vector of filenames of rasters to import.}
}
\value{
Nothing.
}
\description{
GRASS can only deal with raster and vector data in a GRASS mapset. This function takes external rasters and imports them into the current GRASS mapset.
}
