% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterise_stream.R
\name{rasterise_stream}
\alias{rasterise_stream}
\title{Turn a shapefile of stream edges into a raster}
\usage{
rasterise_stream(streams, reference_raster, method, out,
  overwrite = FALSE)
}
\arguments{
\item{streams}{A file name for a shapefile of stream edges or a SpatialLines* object that represents the same thing.}

\item{reference_raster}{File path of a raster layer that defines the extent of the study area, as well as the cell resolution and snapping of the output raster. Can also be a \code{Raster*} object.}

\item{method}{A string. The method that should be used to process the stream. The options are "binary" (stream cells = 1, others = 0) or "unary" (stream cells = NoData, others = 1)}

\item{out}{The filename of the output.}

\item{overwrite}{A logical indicating whether the output is allowed to overwrite existing files. Defaults to \code{FALSE}.}
}
\value{
Nothing. A file will be written to \code{out}. A raster with the name \code{basename(out)} will be written to the current GRASS mapset.
}
\description{
Given a shapefile of lines representing the channels of a stream network, and a reference raster that defines cell resolution and snapping, this function will return a rasterised version of the shapefile.
}
