% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdwplus-package.R
\docType{package}
\name{rdwplus-package}
\alias{rdwplus}
\alias{rdwplus-package}
\title{An Implementation of IDW-PLUS (Inverse Distance Weighted Percent Land Use for Streams) in R}
\description{
Use R to call the hydrological toolboxes and functions in GRASS GIS to compute spatially explicit land-use metrics for stream survey sites. The package includes functions for preprocessing digital elevation and streams data, and one function to compute all the spatially explicit land use metrics described in Peterson et al. (2011) Freshwater Biology, 56(3), 590-610.
}
\author{
\strong{Maintainer}: Alan Pearse \email{alan.pearse@unimelb.edu.au}

Authors:
\itemize{
  \item Grace Heron \email{g.heron@qut.edu.au}
  \item Erin Peterson \email{erin@peterson-consulting.com}
}

}
\keyword{internal}
