% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mapset.R
\name{report_mapset}
\alias{report_mapset}
\title{Identify current mapset or list all possible mapsets}
\usage{
report_mapset(which = "current")
}
\arguments{
\item{which}{One of either 'current' (the default), which causes the function to return the current mapset, or 'possible', which causes the function to list all possible mapsets.}
}
\value{
Nothing.
}
\description{
GRASS GIS uses a system of mapsets.
}
\examples{
\dontrun{

# Get GRASS installation
my_grass <- search_for_grass()[1]

# Create GRASS session
initGRASS(my_grass)

# Ask for current mapset
report_mapset()

# Ask for possible mapsets
report_mapset('possible')

}
}
