% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwd_to_swd.R
\name{rwd_to_swd}
\alias{rwd_to_swd}
\title{Ensure SAGA workspace matches R working directory}
\usage{
rwd_to_swd(...)
}
\arguments{
\item{...}{Additional arguments to \code{rsaga.env}}
}
\value{
A list indicating the current settings for the SAGA environment parameters.
}
\description{
By calling this function, you can set SAGA's working directory to the current working directory you have in R.
}
