% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silence.R
\name{silence}
\alias{silence}
\title{Function to suppress messages, warnings, errors from GRASS commands}
\usage{
silence(value)
}
\arguments{
\item{value}{A logical indicating whether GRASS messages, warnings, errors should be suppressed. Can be missing, and it is missing by default. Choose "TRUE" or "FALSE".}
}
\value{
A logical indicating the current status of the option.
}
\description{
Prevents the printing GRASS warnings, etc. Use with extreme caution. This is only helpful IF AND ONLY IF you are SURE that any printed messages, warnings, and errors are spurious.
}
\examples{
silence(TRUE)
silence(FALSE)
}
