% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle_silence.R
\name{toggle_silence}
\alias{toggle_silence}
\title{Toggle between silence on and silence off}
\usage{
toggle_silence(stay_off = TRUE)
}
\arguments{
\item{stay_off}{A logical indicating whether output suppression should be kept off once it is turned off. That is, if this function is called but output suppression is already off, then for \code{stay_off=TRUE} output suppression will simply remain off. Defaults to \code{TRUE}.}
}
\value{
A logical indicating whether output suppression is active.
}
\description{
This function detects whether output suppression is on or off, and switches it to its opposite state. Under one setting, this function can be used as an off-switch for the GRASS message/warning/error suppression enforced via the use of \code{silence(value = TRUE)}.
}
\examples{
\donttest{ 
\dontrun{ 

## Is silence toggled? 
toggle_silence(T) 

## Is silence NOT toggled?
toggle_silence(F)

}
}
}
