% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burn_in.R
\name{burn_in}
\alias{burn_in}
\title{Burn in streams to a digital elevation model}
\usage{
burn_in(dem, stream, out, burn = 10, overwrite = FALSE,
  max_memory = 300)
}
\arguments{
\item{dem}{A RasterLayer object containing the digital elevation model. This can also be the file path to a raster.}

\item{stream}{A RasterLayer object containing a rasterised stream. The stream pixels should have a value of 1 and non-stream pixels should have a value of 0. This can also be the file path to a raster.}

\item{out}{The filename of the raster to be written out. It will also be imported into the current GRASS mapset with the name \code{basename(out)}.}

\item{burn}{The magnitude of the drainage reinforcement in elevation units. Defaults to \code{10} elevation units.}

\item{overwrite}{A logical indicating whether the file \code{out} should be overwritten in the mapset and on disk. Defaults to \code{FALSE}.}

\item{max_memory}{Max memory used in memory swap mode (MB). Defaults to \code{300}.}
}
\value{
Nothing.
}
\description{
Burning-in streams (also called 'drainage reinforcement') ensures flow direction and accumulation grids based on the digital elevation model will correctly identify the stream path.
}
