% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_iFLS_weights.R
\name{compute_iFLS_weights}
\alias{compute_iFLS_weights}
\title{Compute iFLS weights}
\usage{
compute_iFLS_weights(
  streams,
  null_streams,
  flow_dir,
  out_flow_length,
  out_iFLS,
  out_iFLS_no_stream,
  watershed,
  idwp,
  remove_streams,
  ...
)
}
\arguments{
\item{streams}{Pour point raster containing a single pour point (i.e., the outlet).}

\item{null_streams}{A streams raster with NoData for the stream cells and 1s everywhere else}

\item{flow_dir}{A flow direction raster.}

\item{out_flow_length}{Name of the output flow length raster.}

\item{out_iFLS}{Name of the output weights raster.}

\item{out_iFLS_no_stream}{Name of the output weights raster excluding cells on the stream line (ignored inf \code{remove_streams = FALSE}).}

\item{watershed}{Watershed raster to use as a mask for the flow-path calculations. This is optional for the iFLS weight calculations.}

\item{idwp}{An inverse distance weighting power. This should be negative. The value \code{idwp = -1} is the default.}

\item{remove_streams}{A logical indicating whether cells corresponding to the stream line should be removed from the weights raster. Defaults to \code{FALSE}.}

\item{...}{Optional extra arguments to \code{get_flow_length()}.}
}
\value{
Nothing.
}
\description{
Compute an iFLO weight raster outside of the \code{compute_metrics()} function.
}
