% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{compute_metrics}
\alias{compute_metrics}
\title{Compute spatially explicit watershed land use percentages}
\usage{
compute_metrics(metrics = c("iFLO", "iFLS", "HAiFLO", "HAiFLS"), landuse,
  fields = paste(metrics, landuse, sep = "_"), sites, elevation,
  flow_dir, flow_acc, streams, idwp = -1)
}
\arguments{
\item{metrics}{A character vector. This vector specifies which metric(s) should be calculated. Your options are iFLO, iFLS, iEDO, iEDS, HAiFLO and HAiFLS. The default is to calculate all except for iEDO and iEDS.}

\item{landuse}{Names of landuse or landcover rasters in the current GRASS mapset for which spatially explicit watershed metrics should be computed.}

\item{fields}{A character vector giving the names of the new fields to be created in the sites' attribute table to store the land use metrics. The default is a combination of the metric names and land use raster names.}

\item{sites}{A shapefile of sites; either a file path to the shapefile or a \code{SpatialPoints*} object.}

\item{elevation}{File name of a filled (hydrologically-conditioned) digital elevation model in the current GRASS mapset.}

\item{flow_dir}{A 'Deterministic 8' (D8) flow direction grid derived from \code{derive_flow}.}

\item{flow_acc}{File name of a flow accumulation grid derived from \code{derive_flow} in the current GRASS mapset.}

\item{streams}{File name of a streams raster in the current GRASS mapset. Optional if you are not asking for the iFLS, iEDS, and/or HAiFLS metrics.}

\item{idwp}{The inverse distance weighting parameter. Default is \code{-1}.}
}
\value{
A SpatialPointsDataFrame object, which is the \code{sites} argument with a modified attribute table. The table will contain the new land use metrics.
}
\description{
Workhorse function for rdwplus. This function computes the spatially explicit landuse metrics in IDW-PLUS.
}
