% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{compute_metrics}
\alias{compute_metrics}
\title{Compute spatially explicit watershed land use percentages}
\usage{
compute_metrics(metric = c("iFLO", "iFLS", "HAiFLO", "HAiFLS"), landuse,
  sites, elevation, flow_dir, flow_acc, streams)
}
\arguments{
\item{metric}{A character vector. This vector specifies which metric(s) should be calculated. Your options are iFLO, iFLS, iEDO, iEDS, HAiFLO and HAiFLS. The default is to calculate all except for iEDO and iEDS.}

\item{landuse}{Pointer(s) to land use rasters for which the land use metrics should be computed.}

\item{sites}{Pointer to a shapefile of sites.}

\item{elevation}{A filled (flow-corrected) digital elevation model.}

\item{flow_dir}{A 'Deterministic 8' (D8) flow direction grid derived from \code{elevation}.}

\item{flow_acc}{A flow accumulation grid derived from \code{flow_dir}.}

\item{streams}{A pointer to a streams raster where non-stream cells are coded as 1 and stream cells are coded as NA / NoData. Optional if you are not asking for the iFLS, iEDS, and/or HAiFLS metrics.}
}
\value{
A SpatialPointsDataFrame object, which is the \code{sites} argument with a modified attribute table. The table will contain the new land use metrics.
}
\description{
Workhorse function for rdwplus. This function computes the spatially explicit landuse metrics in IDW-PLUS.
}
