% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_flow.R
\name{derive_flow}
\alias{derive_flow}
\title{Obtain flow direction and accumulation over a digital elevation model (DEM)}
\usage{
derive_flow(dem, flow_dir, flow_acc, flow_tci, sinks, overwrite = FALSE,
  ...)
}
\arguments{
\item{dem}{A digital elevation model that has been hydrologically corrected.}

\item{flow_dir}{The name of the output flow direction file in the current GRASS mapset.}

\item{flow_acc}{The name of the output flow accumulation file in the current GRASS mapset.}

\item{flow_tci}{The name of the output terrain convergence indenx file in the current GRASS mapset.}

\item{sinks}{The name of the output sink-watershed file in the current GRASS mapset.}

\item{overwrite}{Whether any of the outputs should be allowed to overwrite existing files.}

\item{...}{Additional arguments to \code{r.terraflow}.}
}
\value{
Nothing. Files are written in the current GRASS mapset.
}
\description{
This function computes flow direction and accumulation (among other things) from a DEM. This is done using the \code{r.terraflow} tool in GRASS.
}
