% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_streams.R
\name{derive_streams}
\alias{derive_streams}
\title{Extract streams from a flow accumulation raster}
\usage{
derive_streams(flow_acc, out, min_acc = 1000, min_length = 0,
  overwrite = FALSE, ...)
}
\arguments{
\item{flow_acc}{Name of the flow accumulation raster in the current GRASS mapset.}

\item{out}{File path to the output vector dataset of stream lines. Should be WITHOUT .shp extension.}

\item{min_acc}{The minimum accumulation value (in upstream cells) that a cell needs to have in order to be classified as a stream. Defaults to \code{1000}.}

\item{min_length}{The minimum length of a stream segment in cells. Defaults to \code{0}.}

\item{overwrite}{A logical indicating whether the output should be allowed to overwrite existing files. Defaults to \code{FALSE}.}

\item{...}{Additional arguments to \code{r.stream.extract}.}
}
\value{
Nothing. A file with the name \code{paste0(out, ".shp")} will be created and a vector dataset with the name \code{basename(out)} will appear in the current GRASS mapset.
}
\description{
Derive a set of stream lines from a flow accumulation raster.
}
