% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distance.R
\name{get_distance}
\alias{get_distance}
\title{Compute Euclidean distance to a survey site or stream line within a watershed}
\usage{
get_distance(target, out, overwrite = FALSE)
}
\arguments{
\item{target}{File name of the watershed outlet or streams (as a raster) in the current GRASS mapset.}

\item{out}{File path for the result to be written.}

\item{overwrite}{A logical indicating whether the outputs of this function should be allowed to overwrite existing files.}
}
\value{
Nothing. A file with the name \code{basename(out)} will be created in the current GRASS mapset.
}
\description{
This function is needed to compute Euclidean distance from a feature of interest in a watershed raster.
}
