% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flow_length.R
\name{get_flow_length}
\alias{get_flow_length}
\title{Derive a flow length to streams and outlets}
\usage{
get_flow_length(dem, out, overwrite = FALSE)
}
\arguments{
\item{dem}{A file path to the hydrologically corrected DEM, with .sgrd extension.}

\item{out}{A file name for the output raster of flow lengths.}

\item{overwrite}{Overwrite flag. Defaults to \code{FALSE}.}
}
\value{
Nothing. A file with the name \code{out} will be written to SAGA's current workspace.
}
\description{
Given a (hydrologically corrected, see \code{\link{fill_sinks}}) DEM, this function produces a flow accumulation grid which shows the upstream area that flows into each cell in the DEM.
}
