% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flow_length.R
\name{get_flow_length}
\alias{get_flow_length}
\title{Derive a flow length to streams and outlets}
\usage{
get_flow_length(str_rast, flow_dir, out, to_outlet = FALSE,
  overwrite = FALSE, max_memory = 300)
}
\arguments{
\item{str_rast}{Rasterized unary streams file.}

\item{flow_dir}{Flow direction raster.}

\item{out}{A file name for the output raster of flow lengths.}

\item{to_outlet}{Calculate parameters for outlets flag. Defaults to \code{FALSE} for streams.}

\item{overwrite}{Overwrite flag. Defaults to \code{FALSE}.}

\item{max_memory}{Max memory used in memory swap mode (MB). Defaults to \code{300}.}
}
\value{
Nothing. A file with the name \code{out} will be written to GRASS's current workspace.
}
\description{
Given a (hydrologically corrected, see \code{\link{fill_sinks}}) DEM, this function produces a flow accumulation grid which shows the upstream area that flows into each cell in the DEM.
}
