% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_watershed.R
\name{get_watershed}
\alias{get_watershed}
\title{Delineate watershed for a survey site}
\usage{
get_watershed(sites, i, flow_dir, out, write_file = FALSE,
  overwrite = FALSE, max_memory = 300)
}
\arguments{
\item{sites}{A file path to a shapefile of points or a \code{SpatialPoints*} object.}

\item{i}{An integer which indexes one row of the sites' attribute table.}

\item{flow_dir}{The name of a flow direction grid in the current GRASS mapset.}

\item{out}{The name of the output raster. A raster with the name \code{basename(out)} will be imported into the GRASS mapset. If \code{write_file} is true, then a file with the name \code{out} will be written into the user's current working directory.}

\item{write_file}{A logical indicating whether the output file should be stored as a file in the user's current working directory. Defaults to \code{FALSE}.}

\item{overwrite}{A logical indicating whether the output should be allowed to overwrite existing files. Defaults to \code{FALSE}.}

\item{max_memory}{Max memory used in memory swap mode (MB). Defaults to \code{300}.}
}
\value{
Nothing. A raster file with the name \code{out} may be written to the current working directory and one with the name \code{basename(out)} will be imported into the current GRASS mapset.
}
\description{
This function delineates a watershed around the \code{i}th site from a set of survey sites.
}
