% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_watersheds.R
\name{get_watersheds}
\alias{get_watersheds}
\title{Delineate watersheds for survey sites}
\usage{
get_watersheds(sites, flow_dir, out, overwrite = FALSE, lessmem = FALSE)
}
\arguments{
\item{sites}{A file path to a shapefile of points.}

\item{flow_dir}{The name of a flow direction grid in the current GRASS mapset.}

\item{out}{The names of the output watershed rasters.}

\item{overwrite}{A logical indicating whether the output should be allowed to overwrite existing files. Defaults to \code{FALSE}.}

\item{lessmem}{A logical indicating whether to use the less memory modified watershed module. Defaults to \code{FALSE}. If set to \code{TRUE}, the \code{r.wateroutlet.lessmem} extension module must be installed. It can be installed using the GRASS GUI.}
}
\value{
Nothing. A raster file with the name \code{out} may be written to file if you have set the \code{write_file} argument accordingly. A raster with the name \code{basename(out)} will be imported into the current GRASS mapset.
}
\description{
This function delineates watersheds around a set of survey sites.
}
\examples{
# Will only run if GRASS is running
if(check_running()){
}
}
