% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterise_stream.R
\name{rasterise_stream}
\alias{rasterise_stream}
\title{Turn a shapefile of stream edges into a raster}
\usage{
rasterise_stream(streams, out, overwrite = FALSE, max_memory = 300,
  ...)
}
\arguments{
\item{streams}{A file name for a shapefile of stream edges in the current GRASS mapset.}

\item{out}{The filename of the output.}

\item{overwrite}{A logical indicating whether the output is allowed to overwrite existing files. Defaults to \code{FALSE}.}

\item{max_memory}{Max memory used in memory swap mode (MB). Defaults to \code{300}.}

\item{...}{Additional arguments to \code{v.to.rast}.}
}
\value{
Nothing. A file will be written to \code{out}. A raster with the name \code{basename(out)} will be written to the current GRASS mapset.
}
\description{
Given a shapefile of lines representing the channels of a stream network, this function will return a rasterised version of the shapefile. The raster will have the parameters of the current GRASS mapset.
}
