% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterise_stream.R
\name{rasterise_stream}
\alias{rasterise_stream}
\title{Turn a shapefile of stream edges into a raster}
\usage{
rasterise_stream(streams, reference_raster, out, overwrite = FALSE, ...)
}
\arguments{
\item{streams}{A file name for a shapefile of stream edges in the current GRASS mapset.}

\item{reference_raster}{File path of a raster layer in the current GRASS mapset.}

\item{out}{The filename of the output.}

\item{overwrite}{A logical indicating whether the output is allowed to overwrite existing files. Defaults to \code{FALSE}.}

\item{...}{Additional arguments to \code{v.to.rast}.}
}
\value{
Nothing. A file will be written to \code{out}. A raster with the name \code{basename(out)} will be written to the current GRASS mapset.
}
\description{
Given a shapefile of lines representing the channels of a stream network, and a reference raster that defines cell resolution and snapping, this function will return a rasterised version of the shapefile.
}
