% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify_streams.R
\name{reclassify_streams}
\alias{reclassify_streams}
\title{Reclassify streams into the format required for the land use metric calculations}
\usage{
reclassify_streams(stream, out, out_type = "binary", overwrite = FALSE)
}
\arguments{
\item{stream}{Name of a streams raster in the current GRASS mapset. Typically this is the result of \code{rasterise_stream}. The raster should have NoData values for all non-stream cells. Stream cells can have any other value.}

\item{out}{The output file.}

\item{out_type}{Either 'binary', 'unary', or 'none'. See the Description above.}

\item{overwrite}{A logical indicating whether the output should be allowed to overwrite any existing files. Defaults to \code{FALSE}.}
}
\value{
Nothing. A file with the name \code{out} will be written to the current GRASS mapset. This raster will be in unsigned integer format.
}
\description{
Given a streams raster, this function will either create a binary streams raster (0 for non-stream cells and 1 for stream cells) or a unary streams raster (1 for stream cells and NoData for all other cells). Another option is to reclassify the streams raster such that stream cells are given the value NoData and non-stream cells are given the value 1.
}
