% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify_streams.R
\name{reclassify_streams}
\alias{reclassify_streams}
\title{Reclassify streams into the format required for the land use metric calculations.}
\usage{
reclassify_streams(stream, out, out_type = "binary")
}
\arguments{
\item{stream}{A streams raster as either a filepath to a raster or a Raster* object.}

\item{out}{The base name of the output file.}

\item{out_type}{Either 'binary' or 'unary'. See the Description above.}
}
\value{
Nothing. A file with the name \code{out} will be written to the current working directory.
}
\description{
Given a streams raster, this function will either create a binary streams raster (0 for non-stream cells and 1 for stream cells) or a unary streams raster (1 for non-stream cells and NoData for stream cells).
}
