% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_raster.R, R/retrieve_vector.R
\name{retrieve_raster}
\alias{retrieve_raster}
\title{Write a raster layer from the current GRASS mapset to file}
\usage{
retrieve_raster(layer, out_layer, overwrite = FALSE, ...)

retrieve_raster(layer, out_layer, overwrite = FALSE, ...)
}
\arguments{
\item{layer}{The name of the raster in the GRASS mapset that is to be written out.}

\item{out_layer}{The name of the file to be created, with the relevant file extension.}

\item{overwrite}{A logical indicating whether the output from this function should be allowed to overwrite any existing files. Defaults to \code{FALSE}.}

\item{...}{Additional arguments to \code{r.out.gdal}.}

\item{layer}{The name of the vector layer in the GRASS mapset that is to be written out.}

\item{out_layer}{The name of the shapefile to be created (with .shp file extension).}

\item{overwrite}{A logical indicating whether the output from this function should be allowed to overwrite any existing files. Defaults to \code{FALSE}.}

\item{...}{Additional arguments to \code{v.out.ogr}.}
}
\value{
Nothing.

Nothing.
}
\description{
This function writes a GRASS mapset raster to file.

This function writes a GRASS mapset vector layer (like a shapefile) to file.
}
