% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_raster.R
\name{retrieve_raster}
\alias{retrieve_raster}
\title{Write a layer from a GRASS mapset to file}
\usage{
retrieve_raster(layer, out_layer)
}
\arguments{
\item{layer}{The name of the raster in the GRASS mapset that is to be written out.}

\item{out_layer}{The name of the file to be created. If this is a filename only (i.e. no folder is specified in the file path), it will be written to the current working directory.}
}
\value{
Nothing.
}
\description{
This function writes a GRASS mapset raster to file.
}
