% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_raster.R
\name{retrieve_raster}
\alias{retrieve_raster}
\title{Write a raster layer from the current GRASS mapset to file}
\usage{
retrieve_raster(layer, out_layer, overwrite = FALSE, ...)
}
\arguments{
\item{layer}{The name of the raster in the GRASS mapset that is to be written out.}

\item{out_layer}{The name of the file to be created, with the relevant file extension.}

\item{overwrite}{A logical indicating whether the output from this function should be allowed to overwrite any existing files. Defaults to \code{FALSE}.}

\item{...}{Additional arguments to \code{r.out.gdal}.}
}
\value{
Nothing.
}
\description{
This function writes a GRASS mapset raster to file.
}
\examples{
\donttest{
if(!check_running()){
## Initialise session
if(.Platform$OS.type == "windows"){
  my_grass <- "C:/Program Files/GRASS GIS 7.6"
} else {
  my_grass <- "/usr/lib/grass76/"
}
initGRASS(gisBase = my_grass, override = TRUE, mapset = "PERMANENT")

## Load data set
dem <- system.file("extdata", "dem.tif", package = "rdwplus")

set_envir(dem)
raster_to_mapset(rasters = dem, as_integer = FALSE)

## Retrieve raster 
retrieve_raster("dem.tif", out_layer = "retrieved_dem.tif", overwrite = TRUE)

}
}
}
