% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_vector.R
\name{retrieve_vector}
\alias{retrieve_vector}
\title{Write a vector layer from the current GRASS mapset to file}
\usage{
retrieve_vector(layer, out_layer, overwrite = FALSE, ...)
}
\arguments{
\item{layer}{The name of the vector layer in the GRASS mapset that is to be written out.}

\item{out_layer}{The name of the shapefile to be created (with .shp file extension).}

\item{overwrite}{A logical indicating whether the output from this function should be allowed to overwrite any existing files. Defaults to \code{FALSE}.}

\item{...}{Additional arguments to \code{v.out.ogr}.}
}
\value{
Nothing.
}
\description{
This function writes a GRASS mapset vector layer (like a shapefile) to file.
}
\examples{
\donttest{ 
\dontrun{ 

if(!check_running()){
## Initialise session
if(.Platform$OS.type == "windows"){
  my_grass <- "C:/Program Files/GRASS GIS 7.6"
} else {
  my_grass <- "/usr/lib/grass76/"
}
initGRASS(gisBase = my_grass, override = TRUE, mapset = "PERMANENT")

## Load data set
dem <- system.file("extdata", "dem.tif", package = "rdwplus")
stream_shp <- system.file("extdata", "streams.shp", package = "rdwplus")

set_envir(dem)
vector_to_mapset(vectors = stream_shp)

## Retrieve raster 
retrieve_vector("streams", out_layer = "retrieved_streams.shp", overwrite = TRUE)

}
}
}
}
