% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_for_grass.R
\name{search_for_grass}
\alias{search_for_grass}
\title{Find GRASS installations}
\usage{
search_for_grass(guide)
}
\arguments{
\item{guide}{An optional guide folder to search.}
}
\value{
A vector of file paths to potential GRASS installations.
}
\description{
This function finds the path to potential GRASS installations. Warning: this function works by brute force, so it may take a few minutes to find potential GRASS installations.
}
\examples{
\donttest{ 
\dontrun{ 

my_grass <- search_for_grass()
my_grass

}
}
}
