% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_for_grass.R
\name{search_for_grass}
\alias{search_for_grass}
\title{Find GRASS installations}
\usage{
search_for_grass(guide)
}
\arguments{
\item{guide}{Optional. A specific folder to search in for the GRASS installation.}
}
\value{
A vector of file paths to potential GRASS installations.
}
\description{
This function finds the path to potential GRASS installations. It does so in a very crude way; that is, by searching for directories that match the string \code{'GRASS'}.  

Warning: this function works by brute force, so it may take a few minutes to find potential GRASS installations.

Note: This is not guaranteed to work. It is not hard to find the path to your computer's GRASS installation yourself. This is the preferred course of action.
}
\examples{
\donttest{ 

my_grass <- search_for_grass()
my_grass

}
}
