% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mask.R
\name{set_mask}
\alias{set_mask}
\title{Set a raster mask}
\usage{
set_mask(x, inverse = FALSE, overwrite = TRUE, ...)
}
\arguments{
\item{x}{Raster to use as a mask}

\item{inverse}{Whether the inverse of the raster should be used as the mask. Defaults to \code{FALSE}.}

\item{overwrite}{Whether the existing mask should be overwritten. Defaults to \code{TRUE}.}
}
\value{
Nothing.
}
\description{
Set a raster mask
}
