% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap_sites.R
\name{snap_sites}
\alias{snap_sites}
\title{A function to snap sites in a shapefile to a flow accumulation grid}
\usage{
snap_sites(sites, flow_acc, max_move, out, overwrite = FALSE,
  max_memory = 300, use_sp = TRUE, ...)
}
\arguments{
\item{sites}{File name for a shapefile containing the locations of the survey sites in the current GRASS mapset.}

\item{flow_acc}{File name for a flow accumulation raster in the current GRASS mapset.}

\item{max_move}{The maximum distance in cells that any site can be moved to snap it to the flow accumulation grid.}

\item{out}{The output file path.}

\item{overwrite}{Whether the output should be allowed to overwrite any existing files. Defaults to \code{FALSE}.}

\item{max_memory}{Max memory used in memory swap mode (MB). Defaults to \code{300}.}

\item{use_sp}{Logical to use 'sf' or 'stars' classes for vector objects. Defaults to \code{TRUE} to use 'stars' class.}

\item{...}{Additional arguments to \code{r.stream.snap}.}
}
\value{
Nothing. Note that a shapefile of snapped survey sites will be written to the file \code{out} and a shapefile called \code{basename(out)} will be imported into the GRASS mapset.
}
\description{
This function takes a set of survey site locations and makes sure that they are coincident with the point of highest flow accumulation within a specified distance. This is equivalent to snapping sites to a stream network.
}
