% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap_sites.R
\name{snap_sites}
\alias{snap_sites}
\title{A function to snap sites in a shapefile to a stream network}
\usage{
snap_sites(sites, streams, max_move, out, overwrite = FALSE, ...)
}
\arguments{
\item{sites}{File name for a shapefile containing the locations of the survey sites or a \code{SpatialPoints*} object representing the same thing.}

\item{streams}{File name for a shapefile containing the edges of the stream network or a \code{SpatialLines*} object representing the same thing.}

\item{max_move}{The maximum distance in map units that any site can be moved to snap it to the streams.}

\item{out}{The output file path.}

\item{overwrite}{Whether the output should be allowed to overwrite any existing files. Defaults to \code{FALSE}.}

\item{...}{Additional arguments to \code{v.import}.}
}
\value{
Nothing. Note that a shapefile of snapped survey sites will be written to the file \code{out} and a shapefile called \code{basename(out)} will be imported into the GRASS mapset.
}
\description{
This function takes a set of survey site locations and makes sure that they are coincident with the stream network dataset.
}
