% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap_to_flow.R
\name{snap_to_flow}
\alias{snap_to_flow}
\title{A function to snap sites survey sites to a stream raster}
\usage{
snap_to_flow(
  sites,
  flow_acc,
  max_move,
  out,
  overwrite = FALSE,
  max_memory = 300,
  ...
)
}
\arguments{
\item{sites}{File name for a shapefile containing the locations of the survey sites in the current GRASS mapset.}

\item{flow_acc}{Name of a flow accumulation raster in the current GRASS mapset.}

\item{max_move}{The maximum distance in cells that any site can be moved to snap it to the flow accumulation grid.}

\item{out}{Name of the output in the current GRASS mapset}

\item{overwrite}{Whether the output should be allowed to overwrite any existing files. Defaults to \code{FALSE}.}

\item{max_memory}{Max memory (in) used in memory swap mode. Defaults to \code{300} Mb.}

\item{...}{Additional arguments to \code{r.stream.snap}.}
}
\value{
Nothing.
}
\description{
This function takes a set of survey site locations and makes sure that they are coincident with the point nearest the stream line within a specified distance. This is equivalent to snapping sites to a stream network. Note that this function calls \code{r.stream.snap}, which is a GRASS GIS add-on. It can be installed through the GRASS GUI.
}
