% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibe_check.R
\name{vibe_check}
\alias{vibe_check}
\title{A function to summarise the computation region, vectors and rasters in the mapset.}
\usage{
vibe_check()
}
\value{
Nothing.
}
\description{
This function takes no inputs. It prints a list of maps in the current GRASS mapset, as well as the parameters of the current computation region.
}
\examples{
if(check_running()) vibe_check()
}
