% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibe_check.R
\name{vibe_check}
\alias{vibe_check}
\title{A function to print current vectors and rasters in the mapset.}
\usage{
vibe_check()
}
\value{
Nothing.
}
\description{
This function takes no inputs. It prints a list of maps in the current GRASS mapset.
}
\examples{
vibe_check()
}
